pyproject.toml
[project]
name = "secure_stream"
version = "0.0.1"
description = "SecureStream - система мониторинга и удаленного доступа"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "auto-py-to-exe>=2.48.1",
    "ffmpeg-python>=0.2.0",
    "numpy<2",
    "opencv-python==4.8.1.78",
    "pillow>=11.3.0",
    "pyaudio>=0.2.11",
    "pyinstaller>=6.16.0",
    "pyqt6>=6.9.1",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "tk>=0.1.0",
]
